TGV Potentiometer reader rewrite finished
V2 of the TGV potentiometer is working.
This should prevent any skipping of 'clicks' as it detects exactly how much it moved and repeats that to the processing program to be able to get inputs.
This also no longer requires the program to return values real time and can just return values every so often.